1-14 Expand notes field limits. 4-8 Check-in notes - Add notes feature for staff display that are not seen by the public in the OPAC
New Messaging System. This system is indended to replace the old opacnotes and borrowernotes fields in the borrowers table.
This system allows an unlimited number of Libraran and OPAC notes to be attached to a borrower. Each note has a message, a message type,
the data it was created, and which library created it. Each message can only be deleted by the library that created it.

This system may be used simultaneously with the old notes system and does not affect it in any way.

This commit was cherry-picked from git://github.com/kylemhall/koha.git, authored by Kyle Hall

A new database table was added for this feature.