Prepopulate MARC 952 fields when adding additional items to a bib record
I have altered the 'New > New Item' button in the four *detail.pl files
to prepopulate the MARC 952 fields when adding additional items to an
existing bib record.  The exception is the barcode field.

Attempt to fix issue with YUI toolbar in cataloguing.

Cleanup of additem.pl in cataloguing to remove scoping error.

Fixed barcode copy when adding a new item.

Variable assignment was rearranged by the community.  This accounts
for that rearrangement.