1-13 Limit Holds Per Day
Staff has the ability to set a limit on the number of Holds a Borrower can place on items on the same day.
Set using system pref MaxHoldsPerDay

System pref MaxHoldsPerDay is an Integer type pref where the value is the number of holds per day a borrower may place.