1-40 New Items Default Location
NewItemsDefaultLocation. If system pref NewItemsDefaultLocation is set to a location code, all newly catalogued items
will be set to the location set in this preference. Location code must be a valid LOC authorized value type.

The original location code is stored in the new items column 'permanent_location', for future use.