Added syspref LongOverdueToLost
LongOverduetoLost syspref was added to allow staff to assign a certain
number of days after which a overdue item is automatically moved to
lost status.  This is currently set up to run via a cronjog.  The cronjob
will continue to run, but this syspref will override the command line value.