Two sysprefs added to provide a GUI interface for staff to change the number of days to move longoverdues to lost and claimed returns to lost.
Two sysprefs:  LongOverdueToLost and ClaimedReturnToLost have been added.
Although the functionality already existed to some extent with the running
of longoverdue.pl, this allows library staff to change the number of days
to switch the long overdue and claimed return item status to lost.

I would invoke these cronjobs as follows...

5 1 * * * $KOHA_CRON_PATH/longoverdue.pl --lost 30=2 --charge 1 --confirm
7 1 * * * $KOHA_CRON_PATH/longoverdue.pl --lost 30=5 --charge 1 --confirm

So essentially these sysprefs will overwrite the number of 30 days in the
crontab lines above with the values in the syspref.  If these values, are
negative, then this overwrite will not take place.  The crontab lines also
assume you want to charge a replacement fee with the '--charge 1' flag.