Created a reserve expiration date
A new column was added to the reserves (and old_reserves) tables named
expirationdate.  This was done with the following SQL...

  ALTER TABLE reserves ADD expirationdate date;
  ALTER TABLE old_reserves ADD expirationdate date;

This column consists of the hold expiration date which was calculated
from the waitingdate and the number of days in the ReservesMaxPickUpDelay
syspref.

An additional cronjob script called expireholds.pl was also written to
remove holds that have expired from the reserves table and insert them into
the old_reserves table.  It appears that only cancelled and picked up holds
were making it into the old_reserves table and nothing was being done to
account for expired holds.  It's assumed that this would run overnight.