Account for "On hold" display when hold placed upon multiple item bib.
When a hold is placed on a multiple item bib, the previous code would indicate
that all items of that bib were on hold.  This was the case when a next
available hold was placed because a specific item was not tied to that hold.
This code tricks the display into temporarily associating only the current
number of holds versus the number of total items.