Modified SQL query for this report and made changes to inserts into old_reserves
Changes where needed to inserts into old_reserves.  First off, an auto-increment
field called reservenumber was added to the old_reserves table.  Previously,
inserts into old reserves where accomplished by a select * from reserves
table.  This auto-increment variable led to conflicts in reservenumber and,
as a result, the row was never added to old_reserves.  In addition, I had to
take the remove the foreign key associated with items.  This was done by
the following SQL...

ALTER TABLE old_reserves DROP FOREIGN KEY `old_reserves_ibfk_3`;

This needed to be done because when someone would cancel a hold on a "next
vailable" item, an item number would not be assigned (itemnumber=NULL).
The foreign key constraint would then balk at this because there's not a
NULL itemnumber in the items table.